Update URL retrieval method in login callback route
Changed the method of retrieving the URL in both instances within the login callback route. The code now first attempts to retrieve the URL from an environment variable, `NEXT_PUBLIC_HOST`, as a replacement for directly picking it from the request object. This can provide more flexibility and control over the referred URL, especially in diversified deployment scenarios.