Update redirect URL in login callback route
The commit modifies the redirection URL in the login callback route from the origin of the request URL to the application's public host. Yet, if the public host is not available, it will default to the request URL origin. This change is done to ensure proper navigation especially in case of different environments.